requirements.txt
Flask~=2.0.3
numpy
gunicorn
torch~=1.11.0
torchvision~=0.12.0
sklearn
scikit-learn
tqdm
Pillow~=9.0.1
Werkzeug~=2.0.3